+++ source/YaneuraOu.vcxproj.filters
@@ -546,6 +546,9 @@
     <ClCompile Include="eval\deep\nn_types.cpp">
       <Filter>リソース ファイル\eval\deep</Filter>
     </ClCompile>
+    <ClCompile Include="engine\dlshogi-engine\dlshogi_searcher.cpp">
+      <Filter>リソース ファイル\engine\dlshogi-engine</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <Text Include="..\docs\USI拡張コマンド.txt">
