+++ source/engine/dlshogi-engine/UctSearch.h
@@ -23,45 +23,16 @@ namespace dlshogi
 	class UctSearcherGroup
 	{
 	public:
-		UctSearcherGroup() :  threads(0) , gpu_id(-1) , policy_value_batch_maxsize(0) , alloced_policy_value_batch_maxsize(0){}
+		UctSearcherGroup() :  threads(0) , gpu_id(-1) , policy_value_batch_maxsize(0){}
 
 		// 初期化
 		// "isready"に対して呼び出される。
 		// スレッド生成は、やねうら王フレームワーク側で行う。
+		//   model_path                 : 読み込むmodel path
 		//   new_thread                 : このインスタンスが確保するUctSearcherの数
 		//   gpu_id                     : このインスタンスに紐付けられているGPU ID
 		//   policy_value_batch_maxsize : このインスタンスが生成したスレッドがNNのforward()を呼び出す時のbatchsize
-		void Initialize(const int new_thread, const int gpu_id, const int policy_value_batch_maxsize);
-
-		// GPUの初期化
-		// スレッドの初期化が終わっていなければならない。
-		void InitGPU(const std::string& model_path, int gpu_id)
-		{
-			// スレッドが割当たっていないので、GPUの初期化は行う必要がない。
-			if (searchers.size() == 0)
-				return;
-
-			mutex_gpu.lock();
-
-			// やねうら王独自拡張 : モデルpathが異なる時か、batchsizeが異なるかした時だけ生成しなおす。
-			// batchsizeは、forward()するためのbufferをGPU側でメモリを確保しないといけないので、この値が変わった時には
-			// バッファの再確保が必要となる。
-			if (this->model_path != model_path || policy_value_batch_maxsize != alloced_policy_value_batch_maxsize)
-			{
-				// 以前のやつを先に開放しないと次のを確保するメモリが足りないかも知れない。
-				if (nn)
-					nn.reset();
-
-				nn = NN::build_nn(model_path, gpu_id, policy_value_batch_maxsize);
-
-				// 次回、このmodel_pathかalloced_policy_value_batch_maxsizeに変更があれば、再度NNをbuildする。
-				this->model_path = model_path;
-				this->alloced_policy_value_batch_maxsize = policy_value_batch_maxsize;
-			}
-			this->gpu_id = gpu_id;
-
-			mutex_gpu.unlock();
-		}
+		void Initialize(const std::string& model_path , const int new_thread, const int gpu_id, const int policy_value_batch_maxsize);
 
 		// ニューラルネットのforward() (順方向の伝播 = 推論)を呼び出す。
 		void nn_forward(const int batch_size, Eval::dlshogi::NN_Input1* x1, Eval::dlshogi::NN_Input2* x2, Eval::dlshogi::NN_Output_Policy* y1, Eval::dlshogi::NN_Output_Value* y2)
@@ -122,9 +93,6 @@ namespace dlshogi
 		// このインスタンスが生成したスレッドがNNのforward()を呼び出す時のbatchsize
 		// Initialize()で引数として渡される。
 		int policy_value_batch_maxsize;
-
-		// 前回のInitGPU()の時のpolicy_value_batch_maxsizeの値
-		int alloced_policy_value_batch_maxsize;
 
 		// ↑のnnにアクセスする時のmutex
 		std::mutex mutex_gpu;
