+++ source/engine/dlshogi-engine/YaneuraOu_dlshogi_bridge.cpp
@@ -116,7 +116,7 @@ void USI::extra_option(USI::OptionsMap& o)
 #if defined(ONNXRUNTIME)
 	// CPUを使っていることがあるので、default値、ちょっと少なめにしておく。
 	o["DNN_Batch_Size1"]             << USI::Option(32, 1, 1024);
-#else
+#elif defind(TENSOR_RT)
 	// 通常時の推奨128 , 検討の時は推奨256。
 	o["DNN_Batch_Size1"]             << USI::Option(128, 1, 1024);
 #endif
@@ -155,7 +155,7 @@ void Search::clear()
 	searcher.SetResignThreshold((int)Options["Resign_Threshold"]);
 
 	// デバッグ用のメッセージ出力の有無。
-	searcher.debug_message = Options["DebugMessage"];
+	searcher.SetDebugMessage(Options["DebugMessage"]);
 
 	// スレッド数と各GPUのbatchsizeをsearcherに設定する。
 
@@ -175,7 +175,8 @@ void Search::clear()
 		thread_nums.push_back(new_thread[i]);
 		policy_value_batch_maxsizes.push_back(new_policy_value_batch_maxsize[i]);
 	}
-	searcher.SetThread(thread_nums, policy_value_batch_maxsizes);
+
+	searcher.InitGPU(Eval::dlshogi::ModelPaths , thread_nums, policy_value_batch_maxsizes);
 
 	// その他、dlshogiにはあるけど、サポートしないもの。
 
@@ -209,8 +210,6 @@ void Search::clear()
 	searcher.SetPonderingMode(Options["USI_Ponder"]);
 
 	searcher.InitializeUctSearch((NodeCountType)Options["UCT_NodeLimit"]);
-
-	searcher.SetModelPaths(Eval::dlshogi::ModelPaths);
 
 #if 0
 	// dlshogiでは、
